Pull out portions of EigenHelpers that are unrelated to Eigen
Several useful non-eigen functions live in EigenHelpers, which takes a long time to compile
